fix: ensure versionPromise always completes
